feat: Prepare SensorPositionCacheRepo to store type of sensor for specifc position
All sensors have a battery level. The SensorData received for all kind of
sensors is assign to a position, eg. A, B, C, D.
Thoses sensor could be ID: 1 to 12.

The BatteryLevel is also a SensorData and also assign to a position, but
always sensor ID: 13.

To be able to dispatch the battery level to the right vital, we should
keep a cache of last position per sensor type. So that next time we
receive SensorData with sensor 13, we could tell if it:
- TemperatureBatteryLevel
- AirHumidityBatteryLevel
- SoilMoistureBatteryLevel
- LuminosityBatteryLevel